First pass at dealing with different size_t/off_t sizes in C90 environments.
Most of the code dealing with the LogTypes namespace was C which lead to a
lot of nonsensical casting, so I have dumbed LOG_TYPE and LOG_LEVEL down to
plain C even though the move of wiiuse into Source means we don't currently
call GenericLog from C.

Set logging threshold to MAX_LOGLEVEL at startup so debug builds will also
p  rint debugging messages before the GUI is running.

For some reason the way we use SetDefaultStyle doesn't play nice with wx 2.9
so we just get the default black text on a black background. Using a gray
background works around that problem, but I found it to also be much easier
on the eyes so I have switched the background color on all versions.


git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@6528 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/AudioCommon/Src/WaveFile.cpp
@@ -74,7 +74,7 @@ bool WaveFileWriter::Start(const char *filename)
 
 	// We are now at offset 44
 	if (ftello(file) != 44)
-		PanicAlert("wrong offset: %lli", ftello(file));
+		PanicAlert("wrong offset: %lld", (long long)ftello(file));
 
 	return true;
 }
